Fixed bug: for an array section with stride and an empty end-index, e.g. A(1::2) the :: was lexed as tok_double_colon.
Fixed bug: in free form, arithmetic IF, e.g. IF(I-J) 10,20,30 the
first label was lexed as a label due to location where start of a
statement could be (for logical IF), whereas the grammar expects the
labels to be lexed as integer constants.

Fixed bugs: failed to transmit kind parameters of array constructor
using implied-do, logical constant, and character constant.